# testplanit/lib/openapi/custom-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TestPlanIt Custom API",
    "version": "1.0.0",
    "description": "Custom API endpoints for TestPlanIt (non-ZenStack endpoints)"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
...
  "paths": {
...
    "/repository/cases/{caseId}/versions": {
      "post": {
        "tags": ["Repository"],
        "summary": "Record a test case version snapshot",
        "description": "Creates a RepositoryCaseVersions snapshot of the test case's CURRENT state. Any client that changes a case through the model API should call this afterwards, otherwise the case has no version history and a run result recorded against it cannot be read back as the case stood when it ran.\n\nThe caller is responsible for advancing RepositoryCases.currentVersion before calling this, OR for asking this endpoint to do it with `bumpVersion`. Accepts a session cookie or a Bearer API token; a token scoped `mode:read` is rejected with 403.",
        "operationId": "createTestCaseVersion",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID of the test case to snapshot."
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bumpVersion": {
                    "type": "boolean",
                    "description": "Increment the case's currentVersion and snapshot at the new number, in one transaction. Use on EDIT. Mutually exclusive with `version`."
                  },
                  "version": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Explicit version number, for imports preserving history. Mutually exclusive with `bumpVersion`; defaults to the case's currentVersion."
                  },
                  "copyFieldValues": {
                    "type": "boolean",
                    "description": "Copy the case's custom field values onto the version. API clients should set this, otherwise the version renders with every custom field blank. Steps-type fields are never copied."
                  },
                  "creatorId": {
                    "type": "string",
                    "description": "Override the recorded author, for imports. Defaults to the caller."
                  },
                  "creatorName": { "type": "string" },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Override the recorded timestamp, for imports. Defaults to now."
                  },
                  "overrides": {
                    "type": "object",
                    "description": "Values to record on the snapshot instead of the case's current ones. Omit to snapshot the case as stored.",
                    "properties": {
                      "name": { "type": "string" },
                      "stateId": { "type": "integer" },
                      "stateName": { "type": "string" },
                      "automated": { "type": "boolean" },
                      "estimate": { "type": "integer", "nullable": true },
                      "forecastManual": { "type": "integer", "nullable": true },
                      "forecastAutomated": {
                        "type": "number",
                        "nullable": true
                      },
                      "steps": {
                        "type": "array",
                        "items": { "type": "object" },
                        "description": "Tiptap step documents."
                      },
                      "tags": {
                        "type": "array",
                        "items": { "type": "string" }
                      },
                      "issues": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "integer" },
                            "name": { "type": "string" },
                            "externalId": { "type": "string", "nullable": true }
                          }
                        }
                      },
                      "attachments": {
                        "type": "array",
                        "items": { "type": "object" }
                      },
                      "isArchived": { "type": "boolean" },
                      "order": { "type": "integer" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Version snapshot created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "version": {
                      "type": "object",
                      "description": "The created RepositoryCaseVersions row."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid case ID, invalid body, or both `version` and `bumpVersion` given"
          },
          "401": { "description": "No session and no valid API token" },
          "403": { "description": "API token is scoped `mode:read`" },
          "404": {
            "description": "Test case not found, or its project is not accessible to the caller"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth.js session cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      }
    }
  },
  "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }]
}